Issue #2974 - Part 13: Adapt octal number syntax.
Used to be like C, with 0755 or something, now it's
0o755. Also, long and int are the same type now.

2to3 . --fix numliterals -wpn
2to3 . --fix long -wpn

Ref: PEP 3127, PEP 237
https://peps.python.org/pep-3127/
https://peps.python.org/pep-0237/

+++ gfx/gl/GLParseRegistryXML.py
@@ -38,7 +38,7 @@ class GLConstHeader:
     def write(self, arg):
         if isinstance(arg, list):
             self.f.write('\n'.join(arg) + '\n')
-        elif isinstance(arg, (int, long)):
+        elif isinstance(arg, (int, int)):
             self.f.write('\n' * arg)
         else:
             self.f.write(str(arg) + '\n')
