+++ python/PyECC/ecc/primes.py
@@ -35,7 +35,7 @@ def prime(n, k):
         
     for i in range(k):
         
-        a = long(2 + random.randint(0, n - 4))
+        a = int(2 + random.randint(0, n - 4))
         x = exp(a, d, n)
         if (x == 1) or (x == n - 1):
             continue
