+++ python/rsa/rsa/_version133.py
@@ -268,7 +268,7 @@ def extended_euclid_gcd(a, b):
         return (a, 1, 0)
 
     q = abs(a % b)
-    r = long(a / b)
+    r = int(a / b)
     (d, k, l) = extended_euclid_gcd(b, q)
 
     return (d, l, k - l*r)
@@ -331,7 +331,7 @@ def encrypt_int(message, ekey, n):
     n"""
 
     if type(message) is types.IntType:
-        return encrypt_int(long(message), ekey, n)
+        return encrypt_int(int(message), ekey, n)
 
     if not type(message) is types.LongType:
         raise TypeError("You must pass a long or an int")
