+++ testing/mozbase/mozlog/tests/test_logger.py
@@ -80,7 +80,7 @@ class TestStructuredLogging(unittest.TestCase):
         The actual message should contain no fields other than the timestamp
         field and those present in expected."""
 
-        self.assertTrue(isinstance(actual['_time'], (int, long)))
+        self.assertTrue(isinstance(actual['_time'], (int, int)))
 
         for k, v in list(expected.items()):
             self.assertEqual(v, actual[k])
